pyproject.toml
[tool.poetry]
name = "aiozello"
version = "0.1.0"
description = "async Zello Channel API client"
authors = ["Roberto Abdelkader Martínez Pérez <robertomartinezp@gmail.com>"]
license = "Apache"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = ">=3.6.0"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
hypothesis = "^6.82.6"
ipdb = "^0.13.13"
black = "^23.7.0"
ipython = "^8.14.0"
mypy = "^1.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
